+++ src/utils/parser.py
@@ -492,7 +492,9 @@ def build_dfs(archConfigs, filter_metrics, sys_info):
                 if type == "metric_table":
                     headers = ["Metric_ID"]
                     data_source_idx = str(data_config["id"] // 100)
-                    if data_source_idx != 0 or data_source_idx in filter_metrics:
+                    if data_source_idx != 0 or (
+                        filter_metrics and data_source_idx in filter_metrics
+                    ):
                         metric_list[data_source_idx] = panel["title"]
                     if (
                         "cli_style" in data_config