+++ tests/test_utils.py
@@ -25,16 +25,11 @@
 
 import inspect
 import os
+import re
 import shutil
-import subprocess
-from importlib.machinery import SourceFileLoader
 from pathlib import Path
-from unittest.mock import patch
 
 import pandas as pd
-import pytest
-
-rocprof_compute = SourceFileLoader("rocprof-compute", "src/rocprof-compute").load_module()
 
 
 def check_resource_allocation():
@@ -55,6 +50,14 @@ def check_resource_allocation():
                 return
 
     return
+
+
+def check_file_pattern(pattern, file_path):
+    """Check if the given pattern exists in the file"""
+    content = ""
+    with open(file_path) as f:
+        content = f.read()
+    return len(re.findall(pattern, content)) != 0
 
 
 def get_output_dir(suffix="_output", clean_existing=True):
@@ -130,69 +133,3 @@ def check_csv_files(output_dir, num_devices, num_kernels):
         elif file.endswith(".pdf"):
             file_dict[file] = "pdf"
     return file_dict
-
-
-@pytest.fixture
-def binary_handler_profile_rocprof_compute(request):
-    def _handler(config, workload_dir, options=[], check_success=True, roof=False):
-        if request.config.getoption("--call-binary"):
-            baseline_opts = [
-                "build/rocprof-compute.bin",
-                "profile",
-                "-n",
-                "app_1",
-                "-VVV",
-            ]
-            if not roof:
-                baseline_opts.append("--no-roof")
-            process = subprocess.run(
-                baseline_opts
-                + options
-                + ["--path", workload_dir, "--"]
-                + config["app_1"],
-                text=True,
-            )
-            # verify run status
-            if check_success:
-                assert process.returncode == 0
-            return process.returncode
-        else:
-            baseline_opts = ["rocprof-compute", "profile", "-n", "app_1", "-VVV"]
-            if not roof:
-                baseline_opts.append("--no-roof")
-            with pytest.raises(SystemExit) as e:
-                with patch(
-                    "sys.argv",
-                    baseline_opts
-                    + options
-                    + ["--path", workload_dir, "--"]
-                    + config["app_1"],
-                ):
-                    rocprof_compute.main()
-            # verify run status
-            if check_success:
-                assert e.value.code == 0
-            return e.value.code
-
-    return _handler
-
-
-@pytest.fixture
-def binary_handler_analyze_rocprof_compute(request):
-    def _handler(arguments):
-        if request.config.getoption("--call-binary"):
-            process = subprocess.run(
-                ["build/rocprof-compute.bin", *arguments],
-                text=True,
-            )
-            return process.returncode
-        else:
-            with pytest.raises(SystemExit) as e:
-                with patch(
-                    "sys.argv",
-                    ["rocprof-compute", *arguments],
-                ):
-                    rocprof_compute.main()
-            return e.value.code
-
-    return _handler